app/backend/requirements.txt
azure-functions
azure-cosmos>=4.5.0
